2024/bubu CTF/YARARARARARA/requirements.txt
flask
yara-python
